Updated Scene Geometry and Lighting
Added in the scene lighting to create the mood and atmosphere of the
game. Because I was unable to reach the level of darkness that I was
originally planning I added in additional props and closed off the two
hallways. One of the hallways is obscured one with several stacked
barrels.