Added Prototype Restart UI with Toggle Actions
Added a prototype restart UI and game logic to toggle between the start
and restart UI panels. The positioning of the restart panel is not
final and is in its current location for testing purposes. This panel
will be moved to the opposite end of the game scene after user testing
and the camera movement mechanic has been implemented.