Added motion mechanic
Placed the start and restart UI panels in their final locations and
implemented the motion mechanic for moving through the game.

The motion of the camera utilizes a third-party plugin called iTween.
This plugin simplifies the implementation of the motion mechanic. While
this same motion could be accomplished by using a LERP, the plugin
implementation takes three arguments: position, time, and easing —
greatly simplifying the task.

In user testing for the motion experience, both users that were tested
said that the motion was too fast. One person said that the speed felt
unnatural and the other said that it almost made them feel dizzy. At
the time of testing the camera moved from the start UI panel to the
game position in 2 seconds. It took the same amount of time to move
from the game position to the restart UI panel. Following the tests,
and prior to this commit, I decreased the speed so that it now takes 3
seconds to move between each position.