Added Audio and Gameplay Mechanics
Added distinct GvrAudioSources to the indoor and outdoor areas of the
scene.

In addition to audio, I also implemented the gameplay mechanics for the
simon says game. When the player starts the game they move into the
room and the five spheres will light up in a randomly generated
sequence. If the player clicks on the orbs in the correct order they
will continue through the room and out the far door where they are
greeted with the restart UI panel. If the player clicks restart they
are moved back into the play position inside the building and a new
sequence is presented.

Finally, I added the mountains model into the scene to complete the
outdoor area of the game.