Added particle effect when selecting correct sphere
To provide additional positive feedback to the player I have added a
simple particle effect that is instantiated when the player selects the
correct orb in the sequence. If the player selects the wrong orb the
effect is not instantiated and the user will only hear the failure
sound.

To accompany the added particle effect, I have also changed the color
of the orbs on hover to provide better visibility in the scene.