Hide game orbs on win / show game orbs on reset
This update hides the orbs when the player has completed the puzzle.
This prevents the player from moving forward into the sphere and
clipping through it. If the player chooses to play again the camera
moves back into the play position and the orbs reappear.

Just before displaying the spheres again I am calling aestheticReset on
each sphere’s lightUp component to reset its appearance. Before adding
this line of code, when hiding and redisplaying the orbs the last orb
to be selected was still illuminated because the user’s gaze hadn’t
properly left the orb, calling the aestheticReset function.