Tweaked appearance of puzzle orbs and particles
I slightly adjusted the default color of the puzzle spheres so that
they provide greater contrast with the environment. This is in response
to several comments from users who have played the game.

I also adjusted the material for particle system that is displayed when
the orb is correctly selected. I am using a particle material rather
than a standard diffuse shader.